app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.8",
    "jose>=1.0.0",
    "jwcrypto>=1.5.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
